Update `/add_order` to looks for inventyory and cogs accounts
- Needed to pass in the order object from the line_service to the
item_service so that the item can look for it in the both the order
object and config parameter. The method `#find_or_create_by_sku` is used
by multiple endpoints that require line items and I'm passing setting
the payload_object to an empty hash if its not passed in. This should
work but we should be careful in the future in how we use this method;
really need specs to cover the endpoints that use this change.
- This change only affects cases where neew products are created