Bug fix: getting duplicate name error even though the vendor was found
- Previous implmentation was throwing a Duplicate Name Error even though
the name had already been set before. It could possibly be an error with
the sync_token or base service with intuit. It could be trying to create
the vendor instead of updating it. My guess is that the instance found
through `find_by_name` was mutating somehow by the time it was passed to
the `@quickbooks.update`. Since the `#update` method on vendor services
was working, I decided it would be better to catch the duplicate error
and then force it to go through that flow instead
- Reference: https://github.com/ruckus/quickbooks-ruby/issues/164